% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbousuite-package.R
\docType{package}
\name{bbousuite-package}
\alias{bbousuite}
\alias{bbousuite-package}
\title{bbousuite: Easily Install and Load `bbou` (Boreal cariBOU) Packages}
\description{
The `bbou` suite of packages are designed to allow analysis (`bbouretro` and `bboutools`) and simulation (`bbousims`) of Boreal Caribou survival and recruitment data ('bboudata') to estimate population growth rates. A shiny app is also available (`bboushiny`).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://poissonconsulting.github.io/bbousuite/}
}

}
\author{
\strong{Maintainer}: Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID})

Authors:
\itemize{
  \item Seb Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})
  \item John Boulanger \email{boulange@ecological.bc.ca} (\href{https://orcid.org/0000-0001-8222-1445}{ORCID})
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})
  \item Troy Hegel \email{troy.hegel@gov.ab.ca}
  \item Barry Nobert \email{barry.nobert@gov.ab.ca}
  \item Dave Hervieux \email{dave.hervieux@gov.ab.ca}
}

}
\keyword{internal}
