% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bbousuite_packages}
\alias{bbousuite_packages}
\title{List all bbousuite packages}
\usage{
bbousuite_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include bbousuite in the list?}
}
\description{
Inspired by \code{tidyverse::tidyverse_packages()}.
}
\examples{
bbousuite_packages()
}
