% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbousuite-package.R
\docType{package}
\name{bbousuite-package}
\alias{bbousuite}
\alias{bbousuite-package}
\title{bbousuite: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@pissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

}
\keyword{internal}
