% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bbousuite_packages}
\alias{bbousuite_packages}
\title{List all bbousite packages}
\usage{
bbousuite_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include bbousite in the list?}
}
\description{
List all bbousite packages
}
\examples{
bbousuite_packages()
}
